Refactor defs.h, remove heli background
Refactor defs.h and minimize usage of #define. Remove helicopter
background (code + sprite > 1k) to make room for strings and other
more important features.

+++ car_game/src/game/gameobject.h
@@ -14,7 +14,7 @@ class GameObject {
     FP32        xPos;
     FP32        zPos;
     // Screen coordinates
-    int16_t     screenX;
+    int32_t     screenX;
     int16_t     screenY;
     uint8_t     screenW;
